+++ CalculateETA/CalculateETA.csproj
@@ -12,16 +12,17 @@
     <PackageIcon>icon.png</PackageIcon>
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <PackageReleaseNotes>
-      v2.1.2 Methods are sorted.
+      v3.0.0
+      
       See changelog (https://github.com/meokullu/CalculateETA/blob/master/CHANGELOG.md)
     </PackageReleaseNotes>
-    <Version>2.1.1</Version>
+    <Version>3.0.0</Version>
     <Description>CalculateETA is a project to calculate estimated time to arrive on loops whether it is in single-thread or multi-thread applicatons.</Description>
     <Copyright>Enes Okullu</Copyright>
     <RepositoryType>git</RepositoryType>
     <PackageTags>ETA; Time; LeftTime; Iterations; Time-Estimations; Multi-thread; Single-thread; loops</PackageTags>
-    <AssemblyVersion>2.1.2</AssemblyVersion>
-    <FileVersion>2.1.2</FileVersion>
+    <AssemblyVersion>3.0.0</AssemblyVersion>
+    <FileVersion>3.0.0</FileVersion>
     <PackageProjectUrl>https://meokullu.github.io/CalculateETA/</PackageProjectUrl>
     <RepositoryUrl>https://github.com/meokullu/CalculateETA</RepositoryUrl>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
